# goldap/server/docs/swagger.yaml
basePath: /api
definitions:
...
  request.BaseChangePwdReq:
    properties:
      code:
        type: string
      mail:
        maxLength: 100
        minLength: 0
        type: string
    required:
    - code
    - mail
    type: object
  request.BaseSendCodeReq:
    properties:
      mail:
        maxLength: 100
        minLength: 0
        type: string
    required:
    - mail
    type: object
  request.FieldRelationAddReq:
    properties:
      attributes:
        additionalProperties:
          type: string
        type: object
      flag:
        maxLength: 20
        minLength: 1
        type: string
    required:
      - attributes
      - flag
    type: object
  request.FieldRelationDeleteReq:
    properties:
      fieldRelationIds:
        items:
          type: integer
        type: array
    required:
      - fieldRelationIds
    type: object
  request.FieldRelationUpdateReq:
    properties:
      attributes:
        additionalProperties:
          type: string
        type: object
      flag:
        maxLength: 20
        minLength: 1
        type: string
      id:
        type: integer
    required:
      - attributes
      - flag
      - id
    type: object
  request.GroupAddReq:
    properties:
      groupName:
        maxLength: 128
        minLength: 1
        type: string
      groupType:
        maxLength: 20
        minLength: 1
        type: string
      parentId:
        description: 父级Id 大于等于0 必填
        minimum: 0
        type: integer
      remark:
        description: 分组的中文描述
        maxLength: 128
        minLength: 0
        type: string
    required:
    - groupName
    - groupType
    type: object
  request.GroupAddUserReq:
    properties:
      groupId:
        type: integer
      userIds:
        items:
          type: integer
        type: array
    required:
    - groupId
    - userIds
    type: object
  request.GroupDeleteReq:
    properties:
      groupIds:
        items:
          type: integer
        type: array
    required:
    - groupIds
    type: object
  request.GroupRemoveUserReq:
    properties:
      groupId:
        type: integer
      userIds:
        items:
          type: integer
        type: array
    required:
    - groupId
    - userIds
    type: object
  request.GroupUpdateReq:
    properties:
      groupName:
        maxLength: 128
        minLength: 1
        type: string
      id:
        type: integer
      remark:
        description: 分组的中文描述
        maxLength: 128
        minLength: 0
        type: string
    required:
    - groupName
    - id
    type: object
...
  response.ResponseBody:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
host: 127.0.0.1:8150
info:
  contact:
    email: veinstien@163.com
    name: CMPLab-CIMR
    url: https://github.com/cmplab-cimr
  description: 基于Go实现的openLDAP后台管理项目
  termsOfService: https://github.com/cmplab-cimr
  title: Goldap Server
  version: "1.0"
paths:
...
  /base/changePwd:
    post:
      consumes:
      - application/json
      description: 使用邮箱验证码修改密码
      parameters:
      - description: 发送验证码请求数据
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.BaseChangePwdReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      summary: 用户通过邮箱修改密码
      tags:
      - 基础管理
...
  /base/decryptpwd:
    get:
      consumes:
      - application/json
      description: 将加密后的密码解密为明文
      parameters:
      - description: 需要解密的加密密码
        in: query
        name: passwd
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      summary: 密码解密
      tags:
      - 基础管理
  /base/encryptpwd:
    get:
      consumes:
      - application/json
      description: 将明文密码加密
      parameters:
      - description: 需要加密的明文密码
        in: query
        name: passwd
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      summary: 密码加密
      tags:
      - 基础管理
...
  /base/sendcode:
    post:
      consumes:
      - application/json
      description: 向指定邮箱发送验证码
      parameters:
      - description: 发送验证码请求数据
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.BaseSendCodeReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      summary: 发送验证码
      tags:
      - 基础管理
  /fieldrelation/add:
    post:
      consumes:
        - application/json
      parameters:
        - description: 新建字段关系管理记录
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/request.FieldRelationAddReq'
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      security:
        - ApiKeyAuth: [ ]
      summary: 新建字段关系管理记录
      tags:
        - 字段关系管理
  /fieldrelation/delete:
    post:
      consumes:
        - application/json
      parameters:
        - description: 删除字段关系管理记录
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/request.FieldRelationDeleteReq'
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      security:
        - ApiKeyAuth: [ ]
      summary: 删除字段关系管理记录
      tags:
        - 字段关系管理
  /fieldrelation/list:
    get:
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      security:
        - ApiKeyAuth: [ ]
      summary: 获字段关系管理列表
      tags:
        - 字段关系管理
  /fieldrelation/update:
    post:
      consumes:
        - application/json
      parameters:
        - description: 更新字段关系管理记录
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/request.FieldRelationUpdateReq'
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      security:
        - ApiKeyAuth: [ ]
      summary: 更新字段关系管理记录
      tags:
        - 字段关系管理
  /group/add:
    post:
      consumes:
      - application/json
      description: 添加分组记录
      parameters:
      - description: 添加用户记录的结构体
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.GroupAddReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      security:
      - ApiKeyAuth: []
      summary: 添加分组记录
      tags:
      - 分组管理
  /group/adduser:
    post:
      consumes:
      - application/json
      description: 添加用户
      parameters:
      - description: 添加用户记录的结构体
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.GroupAddUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      security:
      - ApiKeyAuth: []
      summary: 添加用户
      tags:
      - 分组管理
  /group/delete:
    post:
      consumes:
      - application/json
      description: 删除分组记录
      parameters:
      - description: 删除用户记录的结构体
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.GroupDeleteReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      security:
      - ApiKeyAuth: []
      summary: 删除分组记录
      tags:
      - 分组管理
  /group/list:
    get:
      consumes:
      - application/json
      description: 获取分组记录列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      security:
      - ApiKeyAuth: []
      summary: 获取分组记录列表
      tags:
      - 分组管理
  /group/removeuser:
    post:
      consumes:
      - application/json
      description: 移除用户
      parameters:
      - description: 移除用户记录的结构体
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.GroupRemoveUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      security:
      - ApiKeyAuth: []
      summary: 移除用户
      tags:
      - 分组管理
...
  /group/syncOpenLdapDepts:
    post:
      consumes:
      - application/json
      description: 同步原ldap部门信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      security:
      - ApiKeyAuth: []
      summary: 同步原ldap部门信息
      tags:
      - 分组管理
  /group/syncSqlGroups:
    post:
      consumes:
      - application/json
      description: 同步Sql中的分组信息到ldap
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      security:
      - ApiKeyAuth: []
      summary: 同步Sql中的分组信息到ldap
      tags:
      - 分组管理
...
  /group/tree:
    get:
      consumes:
      - application/json
      description: 获取分组接口树
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      security:
      - ApiKeyAuth: []
      summary: 获取分组接口树
      tags:
      - 分组管理
  /group/update:
    post:
      consumes:
      - application/json
      description: 更新分组记录
      parameters:
      - description: 更新用户记录的结构体
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.GroupUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      security:
      - ApiKeyAuth: []
      summary: 更新分组记录
      tags:
      - 分组管理
  /group/useringroup:
    get:
      consumes:
      - application/json
      description: 获取分组内用户
      parameters:
      - description: 分组ID
        in: query
        name: groupId
        required: true
        type: integer
      - description: 昵称
        in: query
        name: nickname
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      security:
      - ApiKeyAuth: []
      summary: 获取分组内用户
      tags:
      - 分组管理
  /group/usernoingroup:
    get:
      consumes:
      - application/json
      description: 不在分组的用户
      parameters:
      - description: 分组ID
        in: query
        name: groupId
        required: true
        type: integer
      - description: 昵称
        in: query
        name: nickname
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      security:
      - ApiKeyAuth: []
      summary: 不在分组的用户
      tags:
      - 分组管理
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"